refactor!: deno-first, generate npm
