--- package.json
@@ -1,42 +1,0 @@
-{
-  "name": "csl-next",
-  "version": "0.0.1",
-  "license": "MIT",
-  "type": "module",
-  "scripts": {
-    "build": "node ./build/esbuild.cjs --dev",
-    "build:meta": "node ./build/esbuild.cjs --dev --meta",
-    "build:meta:prod": "node ./build/esbuild.cjs --meta",
-    "build:prod": "node ./build/esbuild.cjs",
-    "lint": "rome check ./src/**.ts",
-    "format": "rome format ./src/**.ts --write .",
-    "format:ci": "rome ci ./src/**.ts",
-    "start": "node ./build/esbuild.cjs --dev --watch --run",
-    "start:ci": "node ./build/esbuild.cjs --run",
-    "start:prod": "node ./build/esbuild.cjs --watch --run",
-    "test": "glob -c \"node --loader tsx --no-warnings --test --test-reporter spec\" \"./tests/**/*.test.ts\""
-  },
-  "devDependencies": {
-    "@es-exec/esbuild-plugin-start": "^0.0.5",
-    "@types/glob": "^8.1.0",
-    "@types/node": "^18.16.3",
-    "class-transformer": "^0.5.1",
-    "esbuild": "^0.17.18",
-    "glob": "^10.2.2",
-    "minimist": "^1.2.8",
-    "reflect-metadata": "^0.1.13",
-    "rome": "^12.0.0",
-    "ts-json-schema-generator": "^1.2.0",
-    "tsx": "^3.12.7",
-    "typedoc": "^0.24.6",
-    "typedoc-umlclass": "^0.7.1",
-    "typescript": "^5.0.4",
-    "yaml-convert": "^1.0.1"
-  },
-  "dependencies": {
-    "ajv": "^8.12.0",
-    "edtf": "^4.4.1",
-    "yaml": "^2.2.2"
-  },
-  "packageManager": "pnpm@8.3.1"
-}
