--- rome.json
@@ -1,12 +1,0 @@
-{
-	"$schema": "./node_modules/rome/configuration_schema.json",
-	"organizeImports": {
-		"enabled": false
-	},
-	"linter": {
-		"enabled": true,
-		"rules": {
-			"recommended": true
-		}
-	}
-}
