+++ src/bibliography.ts
@@ -1,6 +1,4 @@
-import { InputReference, ID } from "./reference.ts";
-import "reflect-metadata";
-import { plainToClass } from "class-transformer";
+import { ID, InputReference } from "./reference.ts";
 
 type BibliographyFile = string; // is there a path type I can use?
