+++ src/cli.ts
@@ -1,8 +1,6 @@
-import { InputBibliography } from "./bibliography.ts";
-import { Style } from "./style.ts";
+import { InputBibliography } from "./lib/bibliography.ts";
 import { Processor } from "./processor.ts";
-//import { Reference } from "./reference";
-import { loadJSON, loadYAML } from "./utils.ts";
+import { loadYAML } from "./lib/utils.ts";
 
 const biby = loadYAML("examples/bibliography.yaml") as InputBibliography;
