+++ src/contributor.ts
@@ -2,13 +2,13 @@
 
 // keep this simple for now
 export class Contributor {
-	constructor(
-		public name: string,
-		public role: string = "author",
-		public parse: boolean = true,
-	) {
-		this.name = name;
-		this.role = role;
-		this.parse = parse;
-	}
+  constructor(
+    public name: string,
+    public role: string = "author",
+    public parse: boolean = true,
+  ) {
+    this.name = name;
+    this.role = role;
+    this.parse = parse;
+  }
 }
