+++ src/reference.ts
@@ -1,7 +1,7 @@
 // Typescript model for a CSL Reference
 
 import { Contributor } from "./contributor.ts";
-import { ReferenceTypes, ContributorRoles } from "./style.ts";
+import { ReferenceTypes } from "./style.ts";
 
 // Types
 
@@ -28,20 +28,20 @@ export type TitleString = string; // plain or Djot?
 // Interfaces
 
 export interface TitleStructured {
-	full?: TitleString;
-	main: TitleString;
-	sub: TitleString[];
+  full?: TitleString;
+  main: TitleString;
+  sub: TitleString[];
 }
 
 /**
  * Input reference data.
  */
 export interface InputReference {
-	type?: ReferenceTypes;
-	title?: Title;
-	author?: Contributor[];
-	editor?: Contributor[];
-	issued?: CSLDate;
-	accessed?: CSLDate;
-	publisher?: string;
+  type?: ReferenceTypes;
+  title?: Title;
+  author?: Contributor[];
+  editor?: Contributor[];
+  issued?: CSLDate;
+  accessed?: CSLDate;
+  publisher?: string;
 }
