+++ src/utils.ts
@@ -1,12 +1,13 @@
 import * as fs from "node:fs";
-import { parse } from "yaml";
+import { parse } from "npm:yaml";
+import { InputBibliography } from "./bibliography.ts";
 
-export function loadJSON(path: string): object {
-	const jdata = fs.readFileSync(path, "utf8");
-	return JSON.parse(jdata);
+export function loadJSON(path: string): InputBibliography {
+  const jdata = fs.readFileSync(path, "utf8");
+  return JSON.parse(jdata);
 }
 
-export function loadYAML(path: string): object {
-	const ydata = fs.readFileSync(path, "utf8");
-	return parse(ydata);
+export function loadYAML(path: string): Record<string, unknown> {
+  const ydata = fs.readFileSync(path, "utf8");
+  return parse(ydata);
 }
